requirements.txt
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
boto3==1.37.1
botocore==1.37.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
fastapi-utils==0.8.0
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.9.0
jmespath==1.0.1
lazy-model==0.2.0
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
openai==1.68.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
postgrest==0.19.3
propcache==0.3.0
psutil==5.9.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.1
redis==5.2.1
requests==2.32.3
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3